notificationapp/requirements.txt
Django==4.1.4
django_environ==0.9.0
djangorestframework==3.14.0
environ==1.0
pytz==2022.6
requests==2.28.1
celery==5.2.7
redis==4.4.0